:lipstick: ValidationMessage - ValidationMessage 파일에 작성하지 않고 index파일에 모든 코드를 넣는걸로 수정
